// packages/grid-table-kanban/src/grid/renderers/cell-renderer/interface.ts
export enum CellType {
  Text = 'Text',
  Link = 'Link',
  Number = 'Number',
  Select = 'Select',
  Image = 'Image',
  Chart = 'Chart',
  Rating = 'Rating',
  User = 'User',
  Boolean = 'Boolean',
  Loading = 'Loading',
  Button = 'Button',
  Date = 'Date',
  Attachment = 'Attachment',
}